Move code to x86::registers
