+++ src/sys/process/spawn.rs
@@ -13,7 +13,8 @@ use super::table::{PROCESS_TABLE, MAX_PROCS};
 use crate::api::process::ExitCode;
 use crate::sys::gdt::GDT;
 use crate::sys::mem;
-use crate::sys::x86::{interrupts, rflags, Cr3};
+use crate::sys::x86::interrupts;
+use crate::sys::x86::registers::{Cr3, flags};
 
 use alloc::boxed::Box;
 use alloc::string::{String, ToString};
@@ -164,7 +165,7 @@ fn exec(ctx: ProcessContext, args_ptr: usize, args_len: usize) {
             "iretq",
             in(reg) GDT.1.user_data.0,
             in(reg) ctx.stack_addr,
-            in(reg) rflags::IF,
+            in(reg) flags::IF,
             in(reg) GDT.1.user_code.0,
             in(reg) ctx.entry_point_addr,
             in("rdi") args_ptr,
